pyproject.toml
[project]
name = "regressor-guided-image-editing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
     "clip",
     "datasets>=4.5.0",
     "diffusers>=0.36.0",
     "kornia>=0.8.2",
     "numpy>=2.4.2",
     "pandas>=3.0.0",
     "pillow>=12.1.1",
     "plot-utils>=0.6.14",
     "pycocotools>=2.0.11",
     "scipy>=1.17.0",
     "torch>=2.10.0",
     "torchvision>=0.25.0",
     "transformers>=5.1.0",
]
[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [ { index="pytorch-cu126" } ]
torchvision = [ { index="pytorch-cu126" } ]
clip = { git = "https://github.com/openai/CLIP.git" }
